pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "robotter-backend-api"
version = "0.1.0"
description = "Robotter Backend API"
authors = ["Mike Henry"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-dotenv = "^1.0.0"
boto3 = "^1.34.0"
python-multipart = "^0.0.6"
PyYAML = "^6.0.1"
fastapi-walletauth = "1.2.3"
solana = "^0.30.2"
numpy = "1.26.4"
robotter-hummingbot = { version = "20241016" }
docker = { git = "https://github.com/felixfontein/docker-py" }
hbot-remote-client = { git = "https://github.com/hummingbot/hbot-remote-client-py.git" }

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.0"
ipdb = "^0.13.13"
flake8 = "^7.0.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^23.12.1"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
